+++ src/endstone/core/level/dimension.h
@@ -28,6 +28,7 @@ public:
     [[nodiscard]] DimensionId getId() const override;
     [[nodiscard]] std::string getTranslationKey() const override;
     [[nodiscard]] Level &getLevel() const override;
+    [[nodiscard]] bool isValid() const override;
     [[nodiscard]] std::unique_ptr<Block> getBlockAt(int x, int y, int z) const override;
     [[nodiscard]] std::unique_ptr<Block> getBlockAt(Location location) const override;
     [[nodiscard]] int getHighestBlockYAt(int x, int z) const override;